+++ components/project-wizard/steps/AreaOfInterestStep.vue
@@ -54,23 +54,11 @@
 
       <div class="project-wizard-area-measurement">
         <span>Area: <strong>{{ formattedArea }}</strong></span>
-        <div
-          class="project-wizard-area-unit-toggle"
-          role="group"
-          aria-label="Area of interest display unit"
-        >
-          <button
-            v-for="option in areaUnitOptions"
-            :key="option.value"
-            class="project-wizard-area-unit-option"
-            :class="{ 'project-wizard-area-unit-option-active': areaDisplayUnit === option.value }"
-            type="button"
-            :aria-pressed="areaDisplayUnit === option.value"
-            @click="emit('update:area-display-unit', option.value)"
-          >
-            {{ option.label }}
-          </button>
-        </div>
+        <area-unit-toggle
+          :model-value="areaDisplayUnit"
+          label="Area of interest display unit"
+          @update:model-value="onAreaDisplayUnitUpdate"
+        />
       </div>
 
       <button
@@ -188,12 +176,8 @@ import resetIcon from '~/assets/img/reset.svg';
 import uploadIcon from '~/assets/img/upload.svg';
 import warningIcon from '~/assets/img/warning.svg';
 
-import {
-  AREA_UNIT_OPTIONS,
-  formatArea,
-  type AreaDisplayUnit,
-} from '~/composables/useAreaDisplayUnit';
 import type { ProjectWizardAreaStepDefinition } from '~/types/project-wizard';
+import { formatArea, type AreaDisplayUnit } from '~/util/area';
 
 interface Props {
   areaDisplayUnit: AreaDisplayUnit;
@@ -217,7 +201,6 @@ const emit = defineEmits<{
 
 const fileInputRef = useTemplateRef<HTMLInputElement>('fileInputRef');
 const isDraggingFile = ref(false);
-const areaUnitOptions = AREA_UNIT_OPTIONS;
 const formattedArea = computed(() =>
   formatArea(props.areaSquareKilometers, props.areaDisplayUnit),
 );
@@ -233,6 +216,10 @@ watch(
 
 function openFilePicker() {
   fileInputRef.value?.click();
+}
+
+function onAreaDisplayUnitUpdate(unit: AreaDisplayUnit) {
+  emit('update:area-display-unit', unit);
 }
 
 function onFileInputChange(event: Event) {
@@ -380,42 +367,6 @@ function onFileDrop(event: DragEvent) {
   gap: 0.65rem;
   color: #195747;
   font-size: 0.88rem;
-}
-
-.project-wizard-area-unit-toggle {
-  display: inline-flex;
-  padding: 0.15rem;
-  background: rgba($text-navy, 0.06);
-  border: 1px solid rgba($text-navy, 0.12);
-  border-radius: 0.45rem;
-}
-
-.project-wizard-area-unit-option {
-  min-width: 2.8rem;
-  padding: 0.3rem 0.45rem;
-  color: $secondary;
-  font-size: 0.78rem;
-  font-weight: 700;
-  line-height: 1;
-  background: transparent;
-  border: 0;
-  border-radius: 0.32rem;
-}
-
-.project-wizard-area-unit-option:hover,
-.project-wizard-area-unit-option:focus-visible {
-  color: $text-navy;
-}
-
-.project-wizard-area-unit-option:focus-visible {
-  outline: 0;
-  box-shadow: 0 0 0 0.18rem rgba($primary, 0.16);
-}
-
-.project-wizard-area-unit-option-active {
-  color: $text-navy;
-  background: #ffffff;
-  box-shadow: 0 0.1rem 0.25rem rgba($text-navy, 0.12);
 }
 
 .project-wizard-area-captured-copy {
